+++ ChangeLog.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 # [Unreleased]
 
+- FIX : Compatibility V16 Dictionnaries *14/06/2022* - 3.14.1
 - NEW : Refonte page setup avec class setup de Dolibarr V15 *11/05/2022* 3.14.0
 - NEW : Ajout de la class TechATM pour l'affichage de la page "A propos" *10/05/2022* 3.13.0
 - NEW : Add total_ht on originproductline tpl if available + add class td identification + add data-id tr identification *07/04/2022* - 3.12.0